Added user local timezone support
