+++ src/screens/homechurch/HomeChurchItem.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import moment from 'moment';
+import moment from 'moment-timezone';
 import {
   View,
   StyleSheet,
@@ -11,6 +11,7 @@ import { Thumbnail } from 'native-base';
 import { Theme, Style } from '../../Theme.style';
 import { HomeChurch } from './HomeChurchScreen';
 import HomeChurchConfirmationModal from './HomeChurchConfirmationModal';
+import { getTimeStamp, getDayOfWeek } from './HomeChurchUtils';
 
 interface Params {
   item: HomeChurch;
@@ -22,47 +23,6 @@ interface Params {
 }
 const { width, height } = Dimensions.get('window');
 
-export const getDayOfWeek = (homechurch: HomeChurch): string => {
-  if (homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly)
-    if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnSunday
-    )
-      return 'Sunday';
-    else if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnMonday
-    )
-      return 'Monday';
-    else if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnTuesday
-    )
-      return 'Tuesday';
-    else if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnWednesday
-    )
-      return 'Wednesday';
-    else if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnThursday
-    )
-      return 'Thursday';
-    else if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnFriday
-    )
-      return 'Friday';
-    else if (
-      homechurch?.schedule?.recurrences?.recurrence?.recurrenceWeekly
-        ?.occurOnSaturday
-    )
-      return 'Saturday';
-    else return moment(homechurch.startDate).format('dddd');
-  else return moment(homechurch?.startDate).format('dddd');
-};
-
 const HomeChurchItem = ({
   active,
   item,
@@ -71,6 +31,7 @@ const HomeChurchItem = ({
   openModal,
   locationToGroupType,
 }: Params): JSX.Element => {
+  const eventTime = getTimeStamp(item);
   const style = StyleSheet.create({
     homeChurchCard: card
       ? {
@@ -170,7 +131,6 @@ const HomeChurchItem = ({
       textDecorationLine: 'underline',
     },
   });
-
   const [confirmationModal, setConfirmationModal] = useState(false);
   const [type, setType] = useState<'contact' | 'calendar' | ''>('');
   return (
@@ -179,7 +139,6 @@ const HomeChurchItem = ({
       {confirmationModal ? (
         <HomeChurchConfirmationModal
           type={type}
-          getDayOfWeek={getDayOfWeek}
           handleClose={() => setConfirmationModal(false)}
           homeChurch={item}
         />
@@ -197,7 +156,10 @@ const HomeChurchItem = ({
           <Text style={style.hmDate}>
             {getDayOfWeek(item)}
             {'\n'}
-            {moment(item?.schedule?.startTime).format('h:mm a')} EDT
+            {moment
+              .tz(getTimeStamp(item), moment.tz.guess())
+              .format('h:mm a')}{' '}
+            {moment.tz(getTimeStamp(item), moment.tz.guess()).format('z')}{' '}
           </Text>
         </View>
         <View>
